feat(auth): add invite code support and registration tracking
- Update `register_verify` endpoint to capture IP address, user agent, and device ID for referral tracking and security.
- Pass `invite_code` from the verification payload to the user registration service.
- Implement `_extract_creem_payer_email` helper to reliably resolve emails from various Creem webhook payload structures.
- Expose billing top-up lookup functions and referral refund processing logic in the router.
- Refine frontend registration form to keep input fields enabled during the verification step.

+++ apps/api/app/schemas/admin_users.py
@@ -8,7 +8,7 @@ class AdminUserItem(BaseModel):
     email: EmailStr
     role: str
     group: str = "default"
-    balance: int
+    balance: float
     banned_at: str | None = Field(default=None, alias="bannedAt")
     created_at: str = Field(alias="createdAt")
     last_login_at: str | None = Field(default=None, alias="lastLoginAt")
@@ -23,7 +23,7 @@ class AdminUsersListResponse(BaseModel):
 
 
 class AdminUserUpdateRequest(BaseModel):
-    balance: int | None = None
+    balance: float | None = None
     banned: bool | None = None
     group: str | None = None
     role: str | None = None
